backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
google-generativeai==0.8.3
websockets==13.1
pydantic==2.9.2
